Arreglo en perfil municipal, actualizacion de las clases adulto y municipal para un mockup inicial en los distintos componentes referentes al apartado web
