+++ AraucariApp/web_app/src/assets/municipal.js
@@ -1,0 +1,16 @@
+/* Estas clases estan codificadas a la fuerza para simular un adulto y un municipal y pasarlo como parametro a los distintos componentes */
+class municipal {
+  constructor() {
+    this.nombre = "Pedro "
+    this.apellido = "Sanchez"
+    this.rut = "12345678-9"
+    this.fnac = "01/01/1980"
+    this.edad = "44"
+    this.direccion = "San Martin 1234"
+    this.sector = "Santa Rosa"
+    this.ciudad = "Temuco"
+    this.photo = "../../avatar.png"
+  }
+}
+
+export default municipal
