--- AraucariApp/web_app/src/components/Dashboard.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function Dashboard () {
-  return (
-    <main className="p-4">
-      <h1 className="text-3xl font-bold">Bienvenido a mi página</h1>
-      <p className="text-lg">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed sit amet nulla auctor, vestibulum magna sed, convallis ex.</p>
-    </main>
-  );
-};
-
-export default Dashboard;
